Long file name from folder fix
