+++ pkg/trace/interpreter/interpreters/open-telemetry/modules/aws/s3_interpret_test.go
