+++ pkg/trace/interpreter/interpreters/open-telemetry/modules/aws/sns_interpret.go
@@ -5,7 +5,7 @@ import (
 	"github.com/StackVista/stackstate-agent/pkg/trace/api"
 	config "github.com/StackVista/stackstate-agent/pkg/trace/interpreter/config"
 	interpreter "github.com/StackVista/stackstate-agent/pkg/trace/interpreter/interpreters"
-	"github.com/StackVista/stackstate-agent/pkg/trace/interpreter/interpreters/open-telemetry/instrumentations"
+	"github.com/StackVista/stackstate-agent/pkg/trace/interpreter/interpreters/open-telemetry/modules"
 	"github.com/StackVista/stackstate-agent/pkg/trace/pb"
 	"github.com/StackVista/stackstate-agent/pkg/util/log"
 	"strings"
@@ -48,7 +48,7 @@ func (t *OpenTelemetrySNSInterpreter) Interpret(spans []*pb.Span) []*pb.Span {
 			var urn = t.CreateServiceURN(strings.ToLower(topicArn))
 			var arn = strings.ToLower(topicArn)
 
-			instrumentations.SpanBuilder(
+			modules.SpanBuilder(
 				span,
 				"consumer",
 				"sns",
