+++ pkg/trace/interpreter/interpreters/open-telemetry/modules/aws/sns_interpret_test.go
