+++ pkg/trace/interpreter/interpreters/open-telemetry/modules/span_builder.go
@@ -1,4 +1,4 @@
-package instrumentations
+package modules
 
 import (
 	"fmt"
