+++ pkg/trace/interpreter/interpreters/open-telemetry/modules/span_builder_test.go
@@ -1,4 +1,4 @@
-package instrumentations
+package modules
 
 import (
 	"github.com/StackVista/stackstate-agent/pkg/trace/pb"
